🧹 Refactor TUI settings screens to use a shared FormScreen base class
🎯 What: Extracted duplicated field navigation logic (tab/shift+tab/up/down) and focus cycling from `ScanSettingsScreen` and `ApiSettingsScreen` into a new `FormScreen` base class in `common.py`.
💡 Why: Improves maintainability by removing 94% similar code blocks across settings screens and centralizes standard form navigation behavior. Also fixes a minor inconsistency where ApiSettingsScreen incorrectly used `action_focus_previous_field` in its bindings.
✅ Verification: Ran the full `uv run pytest` suite successfully. Verified syntax and formatting with `ruff`. Requested automated code review to ensure safety.
✨ Result: `ScanSettingsScreen` and `ApiSettingsScreen` are significantly cleaner, and any future settings screens can simply inherit from `FormScreen` for standard keyboard accessibility.
