+++ src/autoscrapper/tui/common.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from functools import cached_property
 from typing import ClassVar
 
 from textual import events
@@ -7,6 +8,7 @@ from textual.app import ComposeResult
 from textual.binding import Binding
 from textual.containers import Vertical
 from textual.screen import ModalScreen, Screen
+from textual.widget import Widget
 from textual.widgets import Button, Static
 
 
@@ -73,3 +75,55 @@ def update_inline_filter(event: events.Key, text: str) -> tuple[str, bool]:
         return (text + character, True)
 
     return (text, False)
+
+
+class FormScreen(AppScreen):
+    BINDINGS: ClassVar[list[Binding | tuple[str, str] | tuple[str, str, str]]] = [
+        *AppScreen.BINDINGS,
+        Binding("tab", "focus_next_field", "Next field", show=False, priority=True),
+        Binding(
+            "shift+tab",
+            "focus_previous_field",
+            "Previous field",
+            show=False,
+            priority=True,
+        ),
+        Binding(
+            "up",
+            "focus_previous_field",
+            "Previous field",
+            show=False,
+            priority=True,
+        ),
+        Binding("down", "focus_next_field", "Next field", show=False, priority=True),
+    ]
+
+    _FOCUS_ORDER: ClassVar[tuple[str, ...]] = ()
+
+    @cached_property
+    def _cached_focus_widgets(self) -> tuple[Widget, ...]:
+        return tuple(self.query_one(f"#{widget_id}") for widget_id in self._FOCUS_ORDER)
+
+    def _focus_candidates(self) -> list[Widget]:
+        return [w for w in self._cached_focus_widgets if w.is_mounted and not getattr(w, "disabled", False)]
+
+    def _cycle_focus(self, delta: int) -> None:
+        candidates = self._focus_candidates()
+        if not candidates:
+            return
+
+        current = self.focused
+        if current in candidates:
+            index = (candidates.index(current) + delta) % len(candidates)
+        else:
+            index = 0 if delta > 0 else len(candidates) - 1
+
+        target = candidates[index]
+        target.focus()
+        target.scroll_visible(immediate=True)
+
+    def action_focus_next_field(self) -> None:
+        self._cycle_focus(1)
+
+    def action_focus_previous_field(self) -> None:
+        self._cycle_focus(-1)
